requirements.txt
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-Limiter==2.8.2
python-dotenv==0.19.2
requests==2.26.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==9.0.1
cryptography==39.0.1
pywebpush==1.13.0
gunicorn==20.1.0
mysqlclient==2.1.1